исправление ошибок
этап 2, компилируется и хотя бы доходит до линкера
