+++ src/UI/UI.cpp
@@ -4,6 +4,7 @@
 // поэтому при её появлении в заголовочном файле возникают ошибки multiple definition.
 // её использование вряд ли будет где-то дальше этого файла, поэтому оставлю её здесь
 // и не буду пихать инстанс её класса в общий namespace.
+#include "screens.h"
 #include "decibels.h"
 #include "../Hardware/shiftreg.h"
